refactor: extract transform golden test support
