+++ crates/rulemorph/tests/common/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod golden;
 pub(crate) mod trace;
